refactor: Align NewTheme with defaultChild structure
This commit updates the NewTheme plugin to more closely align with the
PKP `defaultChild` theme structure.

Key changes include:
- Renamed plugin file to NewThemePlugin.php (from .inc.php).
- Added namespace APP\plugins\themes\newTheme to NewThemePlugin.php.
- Changed parent theme handle to 'defaultthemeplugin'.
- Updated version.xml to reflect the namespaced class name.
- Bumped plugin version and updated release date.
