+++ newTheme/version.xml
@@ -3,8 +3,8 @@
 <version>
     <application>newTheme</application>
     <type>plugins.themes</type>
-    <release>1.0.0.0</release>
-    <date>2024-07-29</date>
+    <release>1.0.1.0</release> <!-- Bumped patch version for changes -->
+    <date>2025-06-01</date> <!-- Updated date -->
     <lazy-load>0</lazy-load>
-    <class>NewThemePlugin</class>
+    <class>APP\plugins\themes\newTheme\NewThemePlugin</class>
 </version>
